% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-pwr2n.LR.R
\name{cal_event}
\alias{cal_event}
\title{FUNCTION_TITLE}
\usage{
cal_event(ratio, lambda1, lambda0, entry, fup, l_shape, l_scale)
}
\arguments{
\item{ratio}{PARAM_DESCRIPTION}

\item{lambda1}{PARAM_DESCRIPTION}

\item{lambda0}{PARAM_DESCRIPTION}

\item{entry}{PARAM_DESCRIPTION}

\item{fup}{PARAM_DESCRIPTION}

\item{l_shape}{PARAM_DESCRIPTION}

\item{l_scale}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
