% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-evalfup.R
\name{evalfup}
\alias{evalfup}
\title{FUNCTION_TITLE}
\usage{
evalfup(
  object,
  lower.time,
  upper.time,
  size,
  increment = 0.5,
  xlabel = "Follow-up Time",
  ylabl = "Total Sample Size",
  title = "Relationship between Follow-up and \\n Total Sample Size"
)
}
\arguments{
\item{object}{PARAM_DESCRIPTION}

\item{lower.time}{PARAM_DESCRIPTION}

\item{upper.time}{PARAM_DESCRIPTION}

\item{size}{PARAM_DESCRIPTION}

\item{increment}{PARAM_DESCRIPTION, Default: 0.5}

\item{xlabel}{PARAM_DESCRIPTION, Default: 'Follow-up Time'}

\item{ylabl}{PARAM_DESCRIPTION, Default: 'Total Sample Size'}

\item{title}{PARAM_DESCRIPTION, Default: 'Relationship between Follow-up and
Total Sample Size'}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
