% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-gen.wgt.R
\name{gen.wgt}
\alias{gen.wgt}
\title{FUNCTION_TITLE}
\usage{
gen.wgt(method = c("LR"), param, theta)
}
\arguments{
\item{method}{PARAM_DESCRIPTION, Default: c("LR")}

\item{param}{PARAM_DESCRIPTION}

\item{theta}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
