% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-load.mat.R
\name{load.mat}
\alias{load.mat}
\title{Load data matrix for \code{MaxLRtest}}
\usage{
load.mat(dat)
}
\arguments{
\item{dat}{input data set for \code{MaxLRtest} and \code{projection.test}}
}
\value{
the intermediate data set
}
\description{
prepare the intermediate dataset for \code{MaxLRtest} and
\code{projection.test}
}
\seealso{
\code{\link[survival]{survfit}}
\code{\link[zoo]{na.locf}}
}
