% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-n2pwr.maxLR.R
\name{n2pwr.maxLR}
\alias{n2pwr.maxLR}
\title{Power Calculation with Maximum Weighted Logrank Test}
\usage{
n2pwr.maxLR(
  entry = 1,
  fup = 1,
  CtrlHaz,
  hazR,
  transP1,
  transP0,
  Wlist,
  eventN,
  totalN,
  ratio = 1,
  alpha = 0.05,
  alternative = c("two.sided", "less", "greater"),
  k = 100,
  nocensor = FALSE,
  criteria = 100
)
}
\arguments{
\item{entry}{a numeric value indicating the enrollment time, Default: 1}

\item{fup}{a numeric value indicating the minimum follow-up time for subjects.
, Default: 1}

\item{CtrlHaz}{a function,  specifying the hazard function for control group.}

\item{hazR}{a function, specifying the hazard ratio function between
treatment and control group}

\item{transP1}{a numeric vector of length 2, consisting of the transition
probability from
receiving treatment to drop-out (drop-out rate) and
from receiving treatment to receiving control (drop-in rate) per time unit.}

\item{transP0}{a numeric vector of length 2, consisting of the transition
probability from
receiving control to drop-out (drop-out rate) and
from receiving control to receiving treatment (drop-in rate) per time unit.}

\item{Wlist}{a list, consisting of weight functions applied to the test.
The element of the list must be functions. Default is a list of one constant
function, corresponding to the logrank test.}

\item{eventN}{the number of events}

\item{totalN}{the number of subjects}

\item{ratio}{allocation ratoi, Default: 1}

\item{alpha}{type i error, Default: 0.05}

\item{alternative}{alternative hypothesis, Default: c("two.sided", "less", "greater")}

\item{k}{an integer, indicating number of sub-intervals per time unit, Default: 100}

\item{nocensor}{PARAM_DESCRIPTION, Default: FALSE}

\item{criteria}{PARAM_DESCRIPTION, Default: 100}
}
\description{
\code{n2pwr.maxLR} calcualtes the power given either the
number of events or number of subjects
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stats]{weighted.mean}},\code{\link[stats]{cor}}
\code{\link[mvtnorm]{qmvnorm}},\code{\link[mvtnorm]{pmvnorm}}
}
