% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-MaxLRtest.R
\name{plot.MaxLR}
\alias{plot.MaxLR}
\title{FUNCTION_TITLE}
\usage{
\method{plot}{MaxLR}(x, ...)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[graphics]{points}},\code{\link[graphics]{legend}},\code{\link[graphics]{mtext}}
}
