% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-pwr2n.maxLR.R
\name{plot.MaxLRpwr}
\alias{plot.MaxLRpwr}
\title{FUNCTION_TITLE}
\usage{
\method{plot}{MaxLRpwr}(x, type = c("hazard", "survival", "dropout", "event", "censor"), ...)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{type}{PARAM_DESCRIPTION, Default: c("hazard", "survival", "dropout", "event", "censor")}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[graphics]{lines}},\code{\link[graphics]{legend}}
}
