% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-projection.test.R
\name{projection.test}
\alias{projection.test}
\title{FUNCTION_TITLE}
\usage{
projection.test(dat, Wlist, base, alpha = 0.05, side = c("two.sided"))
}
\arguments{
\item{dat}{PARAM_DESCRIPTION}

\item{Wlist}{PARAM_DESCRIPTION}

\item{base}{PARAM_DESCRIPTION}

\item{alpha}{PARAM_DESCRIPTION, Default: 0.05}

\item{side}{PARAM_DESCRIPTION, Default: c("two.sided")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[survival]{survfit}}
\code{\link[MASS]{ginv}}
\code{\link[stats]{Chisquare}}
}
