% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-pwr2n.maxLR.R
\name{pwr2n.maxLR}
\alias{pwr2n.maxLR}
\title{FUNCTION_TITLE}
\usage{
pwr2n.maxLR(
  entry = 1,
  fup = 1,
  CtrlHaz,
  hazR,
  transP1,
  transP0,
  Wlist = list(function(x) {     x^0 }),
  entry_pdf0 = function(x) {     (1/entry) * (x >= 0 & x <= entry) },
  entry_pdf1 = entry_pdf0,
  ratio = 1,
  alpha = 0.05,
  beta = 0.1,
  alternative = c("two.sided"),
  criteria = 500,
  k = 100
)
}
\arguments{
\item{entry}{PARAM_DESCRIPTION, Default: 1}

\item{fup}{PARAM_DESCRIPTION, Default: 1}

\item{CtrlHaz}{PARAM_DESCRIPTION}

\item{hazR}{PARAM_DESCRIPTION}

\item{transP1}{PARAM_DESCRIPTION}

\item{transP0}{PARAM_DESCRIPTION}

\item{Wlist}{PARAM_DESCRIPTION, Default: list(function(x) {
x^0
})}

\item{entry_pdf0}{PARAM_DESCRIPTION, Default: function(x) {
(1/entry) * (x >= 0 & x <= entry)
}}

\item{entry_pdf1}{PARAM_DESCRIPTION, Default: entry_pdf0}

\item{ratio}{PARAM_DESCRIPTION, Default: 1}

\item{alpha}{PARAM_DESCRIPTION, Default: 0.05}

\item{beta}{PARAM_DESCRIPTION, Default: 0.1}

\item{alternative}{PARAM_DESCRIPTION, Default: c("two.sided")}

\item{criteria}{PARAM_DESCRIPTION, Default: 500}

\item{k}{PARAM_DESCRIPTION, Default: 100}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stats]{cor}},\code{\link[stats]{weighted.mean}}
\code{\link[mvtnorm]{qmvnorm}},\code{\link[mvtnorm]{pmvnorm}}
}
