% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-simu.trial.R
\name{simu.trial}
\alias{simu.trial}
\title{FUNCTION_TITLE}
\usage{
simu.trial(
  type = c("event", "time"),
  trial_param,
  bsl_dist = c("weibull", "loglogistic"),
  bsl_param,
  drop_param0,
  drop_param1 = drop_param0,
  entry_pdf0 = function(x) {     (1/trial_param[2]) * (x >= 0 & x <= trial_param[2]) },
  entry_pdf1 = entry_pdf0,
  HR_fun,
  ratio,
  upInt = 100,
  print = TRUE
)
}
\arguments{
\item{type}{PARAM_DESCRIPTION, Default: c("event", "time")}

\item{trial_param}{PARAM_DESCRIPTION}

\item{bsl_dist}{PARAM_DESCRIPTION, Default: c("weibull", "loglogistic")}

\item{bsl_param}{PARAM_DESCRIPTION}

\item{drop_param0}{PARAM_DESCRIPTION}

\item{drop_param1}{PARAM_DESCRIPTION, Default: drop_param0}

\item{entry_pdf0}{PARAM_DESCRIPTION}

\item{entry_pdf1}{PARAM_DESCRIPTION, Default: entry_pdf0}

\item{HR_fun}{PARAM_DESCRIPTION}

\item{ratio}{PARAM_DESCRIPTION}

\item{upInt}{PARAM_DESCRIPTION, Default: 100}

\item{print}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stats]{Weibull}},\code{\link[stats]{integrate}},\code{\link[stats]{Logistic}},\code{\link[stats]{uniroot}},\code{\link[stats]{Uniform}}
}
