% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-load.mat.R
\name{load.mat}
\alias{load.mat}
\title{FUNCTION_TITLE}
\usage{
load.mat(dat)
}
\arguments{
\item{dat}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[survival]{survfit}}
\code{\link[zoo]{na.locf}}
}
