% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-MaxLRtest.R
\name{MaxLRtest}
\alias{MaxLRtest}
\title{FUNCTION_TITLE}
\usage{
MaxLRtest(
  dat,
  Wlist,
  base = c("Combined", "KM", "N"),
  alpha = 0.05,
  side = c("two.sided")
)
}
\arguments{
\item{dat}{PARAM_DESCRIPTION}

\item{Wlist}{PARAM_DESCRIPTION}

\item{base}{PARAM_DESCRIPTION, Default: c("Combined", "KM", "N")}

\item{alpha}{PARAM_DESCRIPTION, Default: 0.05}

\item{side}{PARAM_DESCRIPTION, Default: c("two.sided")}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[survival]{survfit}}
\code{\link[stats]{stepfun}},\code{\link[stats]{cor}}
\code{\link[mvtnorm]{qmvnorm}},\code{\link[mvtnorm]{pmvnorm}}
}
