% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-plotHazSurv.R
\name{plotHazSurv}
\alias{plotHazSurv}
\title{FUNCTION_TITLE}
\usage{
plotHazSurv(
  bsl_dist = c("weibull", "loglogistic"),
  param = c(1.2, 0.03),
  fun_list = fun_list,
  end,
  tit = c("Hazard Function", "Survival Function"),
  pos = c(1, 2),
  f_n = 1
)
}
\arguments{
\item{bsl_dist}{PARAM_DESCRIPTION, Default: c("weibull", "loglogistic")}

\item{param}{PARAM_DESCRIPTION, Default: c(1.2, 0.03)}

\item{fun_list}{PARAM_DESCRIPTION, Default: fun_list}

\item{end}{PARAM_DESCRIPTION}

\item{tit}{PARAM_DESCRIPTION, Default: c("Hazard Function", "Survival Function")}

\item{pos}{PARAM_DESCRIPTION, Default: c(1, 2)}

\item{f_n}{PARAM_DESCRIPTION, Default: 1}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[graphics]{par}},\code{\link[graphics]{lines}},\code{\link[graphics]{mtext}},\code{\link[graphics]{legend}}
\code{\link[stats]{integrate}}
}
