% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-pwr2n.LR.R
\name{pwr2n.LR}
\alias{pwr2n.LR}
\title{FUNCTION_TITLE}
\usage{
pwr2n.LR(
  method = c("schoenfeld", "freedman"),
  lambda0,
  lambda1,
  ratio = 1,
  entry = 0,
  fup,
  alpha = 0.05,
  beta = 0.1,
  alternative = c("two.sided"),
  Lparam = NULL,
  summary = TRUE
)
}
\arguments{
\item{method}{PARAM_DESCRIPTION, Default: c("schoenfeld", "freedman")}

\item{lambda0}{PARAM_DESCRIPTION}

\item{lambda1}{PARAM_DESCRIPTION}

\item{ratio}{PARAM_DESCRIPTION, Default: 1}

\item{entry}{PARAM_DESCRIPTION, Default: 0}

\item{fup}{PARAM_DESCRIPTION}

\item{alpha}{PARAM_DESCRIPTION, Default: 0.05}

\item{beta}{PARAM_DESCRIPTION, Default: 0.1}

\item{alternative}{PARAM_DESCRIPTION, Default: c("two.sided")}

\item{Lparam}{PARAM_DESCRIPTION, Default: NULL}

\item{summary}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stats]{integrate}},\code{\link[stats]{weighted.mean}},\code{\link[stats]{Normal}}
}
