% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-pwr2n.LR.R
\name{pwr2n.LR}
\alias{pwr2n.LR}
\title{FUNCTION_TITLE}
\usage{
pwr2n.LR(
  method = c("schoenfeld", "freedman"),
  lambda0,
  lambda1,
  phi,
  t_enrl = 0,
  t_fup,
  alpha = 0.05,
  beta = 0.1,
  two.side = TRUE
)
}
\arguments{
\item{method}{PARAM_DESCRIPTION, Default: c("schoenfeld", "freedman")}

\item{lambda0}{PARAM_DESCRIPTION}

\item{lambda1}{PARAM_DESCRIPTION}

\item{phi}{PARAM_DESCRIPTION}

\item{t_enrl}{PARAM_DESCRIPTION, Default: 0}

\item{t_fup}{PARAM_DESCRIPTION}

\item{alpha}{PARAM_DESCRIPTION, Default: 0.05}

\item{beta}{PARAM_DESCRIPTION, Default: 0.1}

\item{two.side}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stats]{integrate}},\code{\link[stats]{weighted.mean}},\code{\link[stats]{Normal}}
}
