% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-pwr2n.maxLR.R
\name{pwr2n.maxLR}
\alias{pwr2n.maxLR}
\title{FUNCTION_TITLE}
\usage{
pwr2n.maxLR(
  entry = 1,
  fup = 1,
  k = 100,
  plist1,
  hazR,
  Wlist,
  alpha = 0.05,
  beta = 0.1,
  two.side = TRUE,
  plot = TRUE,
  nocensor = FALSE,
  criteria = 100
)
}
\arguments{
\item{entry}{PARAM_DESCRIPTION, Default: 1}

\item{fup}{PARAM_DESCRIPTION, Default: 1}

\item{k}{PARAM_DESCRIPTION, Default: 100}

\item{plist1}{PARAM_DESCRIPTION}

\item{hazR}{PARAM_DESCRIPTION}

\item{Wlist}{PARAM_DESCRIPTION}

\item{alpha}{PARAM_DESCRIPTION, Default: 0.05}

\item{beta}{PARAM_DESCRIPTION, Default: 0.1}

\item{two.side}{PARAM_DESCRIPTION, Default: TRUE}

\item{plot}{PARAM_DESCRIPTION, Default: TRUE}

\item{nocensor}{PARAM_DESCRIPTION, Default: FALSE}

\item{criteria}{PARAM_DESCRIPTION, Default: 100}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stats]{stepfun}}
\code{\link[mvtnorm]{qmvnorm}},\code{\link[mvtnorm]{pmvnorm}}
\code{\link[graphics]{par}},\code{\link[graphics]{lines}},\code{\link[graphics]{legend}}
}
