% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-simu.fixed.R
\name{simu.fixed}
\alias{simu.fixed}
\title{FUNCTION_TITLE}
\usage{
simu.fixed(
  bsl_dist = c("weibull", "loglogistic"),
  param,
  totN,
  fHR,
  prop,
  T0 = 18,
  Tfup = 24,
  seed,
  upInt = 100
)
}
\arguments{
\item{bsl_dist}{PARAM_DESCRIPTION, Default: c("weibull", "loglogistic")}

\item{param}{PARAM_DESCRIPTION}

\item{totN}{PARAM_DESCRIPTION}

\item{fHR}{PARAM_DESCRIPTION}

\item{prop}{PARAM_DESCRIPTION}

\item{T0}{PARAM_DESCRIPTION, Default: 18}

\item{Tfup}{PARAM_DESCRIPTION, Default: 24}

\item{seed}{PARAM_DESCRIPTION}

\item{upInt}{PARAM_DESCRIPTION, Default: 100}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stats]{Weibull}},\code{\link[stats]{integrate}},\code{\link[stats]{Logistic}},\code{\link[stats]{uniroot}},\code{\link[stats]{Uniform}}
}
