% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-simu.trial.R
\name{simu.trial}
\alias{simu.trial}
\title{Simulate Survival Trial Data}
\usage{
simu.trial(
  type = c("event", "time"),
  trial_param,
  bsl_dist = c("weibull", "loglogistic"),
  bsl_param,
  drop_param,
  entry_pdf = function(x) {     (1/trial_param[2]) * (x >= 0 & x <= trial_param[2]) },
  HR_fun,
  allocation_ratio,
  seed,
  upInt = 100,
  print = TRUE
)
}
\arguments{
\item{type}{indicates whether event-driven trial ("event) or fixed study duration
trial ("time"), Option: c("event", "time")}

\item{trial_param}{a vector of length 3 with components for required subject size, enrollment
time and required number of events ("event" type trial)/follow-up time ("time" type trial)}

\item{bsl_dist}{indicates the survival distribution for control group, option: c("weibull", "loglogistic")}

\item{bsl_param}{a vector of length 2 with the shape and scale parameter for the
survival distribution of control group}

\item{drop_param}{a vector of length 2 with shape and scale parameter for the
weibull distribution of drop-out time}

\item{entry_pdf}{a function describing the pdf of the entry time. The default
is uniform enrollment, Default: function(x) {
(1/trial_param\link{2}) * (x >= 0 & x <= trial_param\link{2})
}}

\item{HR_fun}{a function describing the hazard ratio function between treatment
and control group}

\item{allocation_ratio}{allocation ratio between treatment and control group.
For example, \code{allocation_ratio}=2 if 2:1 allocation is used.}

\item{seed}{a single value, specifying seed for random generation}

\item{upInt}{a value indicating the upper bound usedin the uniroot function.
See details.  Default: 100}

\item{print}{a logical indicating whether basic information summary is printed
to the console or not, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
\code{simu.trial} simulates survival data allowing flexible input
of design parameters. It supports both event-driven design and fixed study duration
design.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stats]{Weibull}},\code{\link[stats]{integrate}},\code{\link[stats]{Logistic}},\code{\link[stats]{uniroot}},\code{\link[stats]{Uniform}}
}
